Protocol: client verification
