--- docs/protocol/client_query_response_verification.go
@@ -1,45 +1,0 @@
-package protocol
-
-func VerifySwarmReceipt(swarmContract SwarmContract, receipt SwarmReceipt) {
-  var minCollateral int64 = 1000000
-
-  // checking that the Swarm node has enough funds
-  var swarmNodeId = receipt.Insurance.PublicKey
-  assertTrue(swarmContract.Deposits[swarmNodeId] >= minCollateral)
-
-  // checking that the receipt is signed by this Swarm node
-  assertTrue(SwarmVerify(receipt.Insurance, receipt.ContentHash))
-}
-
-func VerifyTendermintSignature(flnContract BasicFluenceContract, seal Seal, blockHash Digest) {
-  var minCollateral int64 = 1000000
-
-  // checking that the Tendermint node has enough funds
-  var tmNodeId = seal.PublicKey
-  assertTrue(flnContract.NodesDeposits[tmNodeId] >= minCollateral)
-
-  // checking that the receipt is signed by this Tendermint node
-  assertTrue(TmVerify(seal, blockHash))
-}
-
-func VerifyManifestsReceipts(swarmContract SwarmContract, response QueryResponse) {
-  // checking that manifests and transactions receipts are properly signed by Swarm nodes
-  for _, manifest := range response.Manifests {
-    VerifySwarmReceipt(swarmContract, manifest.LastManifestReceipt)
-    VerifySwarmReceipt(swarmContract, manifest.TxsReceipt)
-  }
-
-  // checking that each manifest points correctly to the previous manifest via the Swarm receipt
-  for i := 0; i < 2; i++ {
-    var manifest = response.Manifests[i + 1]
-    var prevManifest = response.Manifests[i]
-
-    assertEq(manifest.LastManifestReceipt.ContentHash, SwarmHash(pack(prevManifest)))
-  }
-}
-
-func VerifyResponseChunks(results QueryResponse) {
-  for k := range results.Chunks {
-    assertTrue(VerifyMerkleProof(results.Chunks[k], results.Proofs[k], results.Manifests[0].VMStateHash))
-  }
-}
